refactor: remove utils dir
